Implement a dashboard
Implement a dashboard showing all tetromino shapes in the queue,
score and the number of cleared lines.

Refactor the parameter list of GridGraphic constructor.

In Tetromino class, add an additional method to change tetrominos
position in the logical grid. This method is supposed to be used
only to set tetrominos in the dashboard. For game usage, the move
method shall be used instead,

In ShapeI class, implement an additional method to retrieve the shapes type.
This information is then used to insert the right tetromino shape into
the dashboard queue. Furthermore, provide the initial position in the ShapeI
constructor as an default argument to decouple the placement of the shapes
between the two grids (normal grid and dashboard grid).

In TetrominoGraphic class, propagate the retrieving tetromino type and tetromino
position in logical grid frame.

Clean up Tetromino tests.