Game over notification
In PlayGround class, implement the recognition
when the game is over, notify the player about
occurring game over and how to restart the game
via a shortcut (Ctrl + N). Furthermore, implement
an additional method to start a new game.

In Dashboard class, implement a method to reset
the scoring and delete the queue with the next
elements.

In GridLogic class, implement a method to free
entire grid at once. This method is then used to
reset all states when the game is restarted.

In Game class, stop performing periodic shapes
dropping and drawing when game is over.

In TetrominoGraphic class, implement an additional
method retrieving the highest row of a given shape.